Fix on css model (angular client) , moved schema-crawler dependency to a helper class.
