+++ src/main/java/org/nanotek/crawler/data/util/db/JdbcHelper.java
@@ -507,7 +507,7 @@ public class JdbcHelper {
 		.forEach(f ->{
 			processForeignKey(f , meta);
 		});
-		meta.setTable(t);
+		meta.getRdbmsClass().setTable(t);
 		meta.setClassName(t.getFullName());
 		String newName = processNameTranslationStrategy(t.getName());
 		meta.setClassName(newName);
@@ -553,12 +553,12 @@ public class JdbcHelper {
 		meta.getMetaRelationsClasses()
 		.stream()
 		.forEach(c -> {
-			if(c.getPrimaryKeyTable().equals(meta.getTable())) {
+			if(c.getPrimaryKeyTable().equals(meta.getRdbmsClass(). getTable())) {
 				c.getColumns()
 				.stream()
 				.forEach(cc ->{
 					if (cc.getName().equalsIgnoreCase(md.getColumnName())) {
-						meta.addReferencedTable(c.getReferencedTable());
+						meta.getRdbmsClass(). addReferencedTable(c.getReferencedTable());
 					}
 				});
 			}
